introduce routing module, move routing related pieces to new lib layout
