+++ projects/storefrontlib/src/cms-components/checkout/checkout-config.service.spec.ts
@@ -2,7 +2,7 @@ import { TestBed } from '@angular/core/testing';
 import { ActivatedRoute } from '@angular/router';
 import { RoutesConfig, RoutingConfigService } from '@spartacus/core';
 import { of } from 'rxjs';
-import { defaultStorefrontRoutesConfig } from '../../lib/ui/pages/default-routing-config';
+import { defaultStorefrontRoutesConfig } from '../../cms-structure/routing/default-routing-config';
 import { CheckoutConfigService } from './checkout-config.service';
 import { CheckoutConfig } from './config/checkout-config';
 import { defaultCheckoutConfig } from './config/default-checkout-config';
