+++ projects/storefrontlib/src/cms-components/checkout/components/checkout-progress/checkout-progress.component.spec.ts
@@ -11,7 +11,7 @@ import {
   RoutingService,
 } from '@spartacus/core';
 import { Observable, of } from 'rxjs';
-import { defaultStorefrontRoutesConfig } from '../../../../lib/ui/pages/default-routing-config';
+import { defaultStorefrontRoutesConfig } from '../../../../cms-structure/routing/default-routing-config';
 import { CheckoutConfig } from '../../config/checkout-config';
 import { defaultCheckoutConfig } from '../../config/default-checkout-config';
 import { CheckoutProgressComponent } from './checkout-progress.component';
