+++ projects/storefrontlib/src/cms-components/checkout/guards/checkout.guard.spec.ts
@@ -1,11 +1,10 @@
 import { TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { RoutesConfig, RoutingConfigService } from '@spartacus/core';
-import { defaultStorefrontRoutesConfig } from '../../../lib/ui/pages/default-routing-config';
+import { defaultStorefrontRoutesConfig } from '../../../cms-structure/routing/default-routing-config';
 import { CheckoutConfig } from '../config/checkout-config';
 import { defaultCheckoutConfig } from '../config/default-checkout-config';
 import { CheckoutGuard } from './checkout.guard';
-
 
 const MockCheckoutConfig: CheckoutConfig = defaultCheckoutConfig;
 const MockRoutesConfig: RoutesConfig = defaultStorefrontRoutesConfig;
