+++ projects/storefrontlib/src/cms-components/checkout/guards/delivery-mode-set.guard.spec.ts
@@ -7,7 +7,7 @@ import {
   ServerConfig,
 } from '@spartacus/core';
 import { Observable, of } from 'rxjs';
-import { defaultStorefrontRoutesConfig } from '../../../lib/ui/pages/default-routing-config';
+import { defaultStorefrontRoutesConfig } from '../../../cms-structure/routing/default-routing-config';
 import { CheckoutConfigService } from '../checkout-config.service';
 import { CheckoutConfig } from '../config/checkout-config';
 import { defaultCheckoutConfig } from '../config/default-checkout-config';
